Decouple the TS and Rust stacks: retire drive mode, lockstep gates, and handle FFI
The stacks were verified move-for-move (and 3,000 games outcome-identical)
before this cut, so the verification machinery has done its job. Rust is now
the free-running prototyping substrate; port-backs to the game's TS CPU mode
carry no bit-identicality requirement. The only remaining runtime seam is
chomp_run_games: bun serializes teams + the address book, Rust plays whole
games natively.

Retired (all in git history if parity ever needs re-proving):
- sims/src/arena/rust-engine.ts — the drive-mode adapter (TS strategies
  reading the Rust engine through the handle API)
- transpiler/scripts/arena_rust_lockstep.ts — drive-mode move-for-move gate
- transpiler/scripts/strategy_lockstep.ts — batch-mode per-turn equality gate
- transpiler/scripts/arena_benchmark.ts — the TS-reference recording run
  (and the winrate-compare flow with it, per request)
- ffi-rs handle API: Session registry, snapshot/rich-state serializers,
  per-battle handles — lib.rs is now batch-only (chomp_ffi_version 0x5,
  asserted at dlopen in the new slim sims/src/arena/rust-ffi.ts bridge)
- --engine flag threading in mon-data*.ts, __runHypotheticalFork/__disposeFork
  delegation in transpose.ts/forward-model.ts, Sim::global_kv

batch_benchmark.ts is now a pure benchmark of the native stack (no recorded
TS reference). Doc perf figures widened to honest ranges (~230-330 games/s
1-thread, ~880-1,270 4-thread): a regression scare during gating bisected to
box variance — HEAD code measured 229 games/s in the same window where it
had measured 329 earlier; the decoupling diff itself is perf-neutral.

Gates: cargo build/test green; batch_benchmark 3,000 games 0 errors
(221/890 games/s at 1/4 threads this window); TS arena mon-data smoke runs
clean on the slimmed game loop.
