+++ sims/src/arena/mon-data-worker.ts
@@ -8,7 +8,7 @@ import { runItems, type ShardResult, type WorkItem } from './mon-data-core';
 declare var self: Worker;
 
 self.onmessage = (e: MessageEvent) => {
-  const { items, maxTurns, engine } = e.data as { items: WorkItem[]; maxTurns: number; engine?: 'ts' | 'rust' };
-  const res: ShardResult = runItems(items, maxTurns, engine ?? 'ts');
+  const { items, maxTurns } = e.data as { items: WorkItem[]; maxTurns: number };
+  const res: ShardResult = runItems(items, maxTurns);
   postMessage(res);
 };
